Make `anchorPoint` a `UnitPoint` instead of `CGPoint`
`UnitPoint` - https://developer.apple.com/documentation/swiftui/unitpoint - is more SwiftUI-y than `CGPoint`, and will allow the anchorPoint to use the constants like `.center` and `.bottom`.

Extract extensions to make it trivial to convert the `UnitPoint` to a `centerOffset` within a rect, and to convert back to a `CGPoint` for the `UIView` `anchorPoint` property.